requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
validators==0.20.0
requests==2.31.0
